Datagrid: change bools only when clicked
Previously, any checkbox in a row would toggle when the row was clicked
This is a bit dangerous. All usages of booleans we've seen represent
membership to a group and they are usually important. Accidentally
changing them is bad.

Now, you have to click on the checkbox cell to change it, not just
anywhere on the row.